feat: update PlaywrightWebDriver and related classes for improved configuration and selector utilities
