+++ src/main/java/io/github/kmariusz/playwrightwebdriver/PlaywrightWebDriver.java
@@ -19,9 +19,21 @@ import com.microsoft.playwright.BrowserContext;
 import com.microsoft.playwright.Page;
 import com.microsoft.playwright.Playwright;
 
+import io.github.kmariusz.playwrightwebdriver.config.PlaywrightWebDriverOptions;
+import io.github.kmariusz.playwrightwebdriver.util.JavaScriptUtils;
+import io.github.kmariusz.playwrightwebdriver.util.SelectorUtils;
+
 /**
  * A Selenium WebDriver implementation that uses Playwright as the underlying automation engine.
  * This class extends RemoteWebDriver to provide Playwright's capabilities within the Selenium API.
+ * <p>
+ * PlaywrightWebDriver enables using Playwright's modern automation capabilities while maintaining
+ * compatibility with the Selenium WebDriver interface. This allows existing Selenium-based test
+ * suites to leverage Playwright's performance and reliability advantages with minimal code changes.
+ * <p>
+ * Note that some advanced Playwright features may not be directly accessible through the standard 
+ * WebDriver interfaces. For these cases, use the {@link #getPlaywrightPage()} and 
+ * {@link #getPlaywrightContext()} methods to access the native Playwright objects.
  */
 public class PlaywrightWebDriver extends RemoteWebDriver {
     /** The Playwright instance used for browser automation. */
@@ -36,10 +48,10 @@ public class PlaywrightWebDriver extends RemoteWebDriver {
     /** The Page instance representing a single tab or window within the browser. */
     private final Page page;
     
-    /** Map of window handles to their URLs, used for window handle management. */
+    /** Map of window handles (UUIDs) to their URLs, for multi-window session management. */
     private final Map<String, String> windowHandles = new HashMap<>();
     
-    /** The current window handle identifier. */
+    /** The identifier for the currently active window handle. */
     private String currentWindowHandle;
 
     /**
@@ -52,18 +64,18 @@ public class PlaywrightWebDriver extends RemoteWebDriver {
     /**
      * Creates a new PlaywrightWebDriver instance with the specified options.
      *
-     * @param options the configuration options for this WebDriver instance
+     * @param options the configuration options for this WebDriver instance, or null for default options
+     * @throws IllegalArgumentException if an unsupported browser type is specified in the options
+     * @throws RuntimeException if there's a failure initializing Playwright components
      */
     public PlaywrightWebDriver(PlaywrightWebDriverOptions options) {
         if (options == null) {
             options = new PlaywrightWebDriverOptions();
         }
 
-        this.playwright = Playwright.create();
+        this.playwright = Playwright.create(options.getCreateOptions());
         this.browser = createBrowser(options);
-        this.context = browser.newContext(new Browser.NewContextOptions()
-                .setViewportSize(options.getWindowWidth(), options.getWindowHeight())
-                .setIgnoreHTTPSErrors(options.isIgnoreHTTPSErrors()));
+        this.context = browser.newContext(options.getContextOptions());
         this.page = context.newPage();
         this.currentWindowHandle = UUID.randomUUID().toString();
 
@@ -86,20 +98,16 @@ public class PlaywrightWebDriver extends RemoteWebDriver {
      *
      * @param options the configuration options for the browser
      * @return a configured Browser instance
-     * @throws IllegalArgumentException if the browser type is not supported
+     * @throws IllegalArgumentException if the browser type specified in options is not supported
      */
     private Browser createBrowser(PlaywrightWebDriverOptions options) {
-        com.microsoft.playwright.BrowserType.LaunchOptions launchOptions = new com.microsoft.playwright.BrowserType.LaunchOptions()
-                .setHeadless(options.isHeadless())
-                .setSlowMo(options.getSlowMo());
-
         switch (options.getBrowserType()) {
             case CHROMIUM:
-                return playwright.chromium().launch(launchOptions);
+                return playwright.chromium().launch(options.getLaunchOptions());
             case FIREFOX:
-                return playwright.firefox().launch(launchOptions);
+                return playwright.firefox().launch(options.getLaunchOptions());
             case WEBKIT:
-                return playwright.webkit().launch(launchOptions);
+                return playwright.webkit().launch(options.getLaunchOptions());
             default:
                 throw new IllegalArgumentException("Unsupported browser type: " + options.getBrowserType());
         }
@@ -137,6 +145,7 @@ public class PlaywrightWebDriver extends RemoteWebDriver {
 
     /**
      * Finds all elements within the current page that match the given search criteria.
+     * The search criteria are converted to Playwright-compatible selectors.
      *
      * @param by the locating mechanism to use
      * @return a list of WebElements matching the search criteria
@@ -153,6 +162,7 @@ public class PlaywrightWebDriver extends RemoteWebDriver {
 
     /**
      * Finds the first element within the current page that matches the given search criteria.
+     * The search criteria are converted to Playwright-compatible selectors.
      *
      * @param by the locating mechanism to use
      * @return the first WebElement matching the search criteria
@@ -167,7 +177,7 @@ public class PlaywrightWebDriver extends RemoteWebDriver {
     /**
      * Gets the source code of the current page.
      *
-     * @return the source code of the current page as a string
+     * @return the HTML source code of the current page as a string
      */
     @Override
     public String getPageSource() {
@@ -176,6 +186,7 @@ public class PlaywrightWebDriver extends RemoteWebDriver {
 
     /**
      * Closes the current browser session and releases all associated resources.
+     * This method closes the page, context, browser, and playwright instances in sequence.
      */
     @Override
     public void close() {
@@ -195,6 +206,7 @@ public class PlaywrightWebDriver extends RemoteWebDriver {
 
     /**
      * Quits the driver, closing all associated windows/tabs and releasing resources.
+     * This is equivalent to calling {@link #close()} in this implementation.
      */
     @Override
     public void quit() {
@@ -204,7 +216,7 @@ public class PlaywrightWebDriver extends RemoteWebDriver {
     /**
      * Gets the set of window handles available to the driver.
      *
-     * @return a set of window handle identifiers
+     * @return a set of window handle identifiers for all open windows/tabs
      */
     @Override
     public Set<String> getWindowHandles() {
@@ -214,7 +226,7 @@ public class PlaywrightWebDriver extends RemoteWebDriver {
     /**
      * Gets the handle of the current window.
      *
-     * @return the window handle identifier for the current window
+     * @return the window handle identifier for the currently active window
      */
     @Override
     public String getWindowHandle() {
@@ -226,7 +238,7 @@ public class PlaywrightWebDriver extends RemoteWebDriver {
      *
      * @param script the JavaScript code to execute
      * @param args the arguments to pass to the script
-     * @return the value returned by the script
+     * @return the value returned by the script, or null if the script returns undefined
      */
     @Override
     public Object executeScript(String script, Object... args) {
@@ -239,7 +251,7 @@ public class PlaywrightWebDriver extends RemoteWebDriver {
      *
      * @param script the JavaScript code to execute asynchronously
      * @param args the arguments to pass to the script
-     * @return the value returned by the script
+     * @return the value returned by the script when the Promise resolves
      */
     @Override
     public Object executeAsyncScript(String script, Object... args) {
@@ -252,7 +264,7 @@ public class PlaywrightWebDriver extends RemoteWebDriver {
      * @param target the output type for the screenshot
      * @param <X> the return type of the screenshot output
      * @return the screenshot as the specified output type
-     * @throws WebDriverException if the screenshot could not be taken
+     * @throws WebDriverException if the screenshot could not be taken or processed
      */
     @Override
     public <X> X getScreenshotAs(OutputType<X> target) throws WebDriverException {
@@ -263,6 +275,7 @@ public class PlaywrightWebDriver extends RemoteWebDriver {
 
     /**
      * Gets the Playwright Page instance used by this driver.
+     * This provides direct access to Playwright-specific functionality not available through the WebDriver interface.
      *
      * @return the Playwright Page instance
      */
@@ -272,6 +285,7 @@ public class PlaywrightWebDriver extends RemoteWebDriver {
 
     /**
      * Gets the Playwright BrowserContext instance used by this driver.
+     * This provides direct access to context-level Playwright functionality such as cookies, permissions, and network control.
      *
      * @return the Playwright BrowserContext instance
      */
